Update README.md
I want to show some of my programming activities demanded by my university. This is a form to show myself how i am evolving in this area and to make clear to other people or companies interested in my services about my current skills and capacities, even though i am still in an initial level of learning and i still need to learn a lot to be a good professional in the I.T area.

My name is Patrick, i am twenty four years old and i am from Brazil. One of my dreams is to become a renowned professional in the I.T area, and i will achieve it!